Break up http_server into additional packages
