+++ src/main/java/directory_page_creator/DirectoryPageCreator.java
@@ -1,19 +1,21 @@
-package http_server;
+package directory_page_creator;
 
 import file_handler.FileHandler;
+import html_builder.HTMLBuilder;
+
 import java.util.List;
 
-public class DirectoryCreator {
+public class DirectoryPageCreator {
     private List<String> directoryList;
-    private String newStaticDirectoryPath;
+    private String staticDirectoryPath;
 
-    public DirectoryCreator(List<String> directoryList, String newStaticDirectoryPath) {
+    public DirectoryPageCreator(List<String> directoryList, String staticDirectoryPath) {
         this.directoryList = directoryList;
-        this.newStaticDirectoryPath = newStaticDirectoryPath;
+        this.staticDirectoryPath = staticDirectoryPath;
     }
 
     public String generateHTML() {
-        String directoryCSSPath = "./src/main/java/http_server/DirectoryStyles";
+        String directoryCSSPath = "./src/main/java/directory_page_creator/DirectoryStyles";
         HTMLBuilder builder = new HTMLBuilder();
 
         return builder
@@ -21,7 +23,7 @@ public class DirectoryCreator {
             .addHeader(FileHandler.getFileContents(directoryCSSPath))
             .addHeader("</style>")
             .append("<div class='directory-page'>")
-            .append("<h1>Directory for " + newStaticDirectoryPath + "</h1>")
+            .append("<h1>Directory for " + staticDirectoryPath + "</h1>")
             .append("<hr />")
             .append("<ul>\n")
             .append(renderAsList())
@@ -33,7 +35,7 @@ public class DirectoryCreator {
     private String renderAsList() {
         String htmlList = "";
         for (int i = 0; i < directoryList.size(); i++) {
-            htmlList += "<li class='bullets'><a href='" + newStaticDirectoryPath + "/" +
+            htmlList += "<li class='bullets'><a href='" + staticDirectoryPath + "/" +
                     directoryList.get(i) + "'>" + directoryList.get(i) + "</a></li>\n";
         }
         return htmlList;
