+++ src/test/java/directory_page_creator/DirectoryPageCreatorTest.java
@@ -1,4 +1,4 @@
-package http_server;
+package directory_page_creator;
 
 import file_handler.FileHandler;
 import org.junit.jupiter.api.AfterEach;
@@ -7,7 +7,7 @@ import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import java.util.List;
 
-public class DirectoryCreatorTest {
+public class DirectoryPageCreatorTest {
     @BeforeEach
     public void prepFiles() {
         FileHandler.deleteDirectory("./public/dog");
@@ -24,7 +24,7 @@ public class DirectoryCreatorTest {
     public void static_files_in_public_directory_are_rendered_as_a_HTML_page_with_contents_listed() {
         String path = "./public";
         List<String> directoryContents = new FileHandler().readDirectoryContents(path);
-        DirectoryCreator directoryCreator = new DirectoryCreator(directoryContents, "/public");
+        DirectoryPageCreator directoryPageCreator = new DirectoryPageCreator(directoryContents, "/public");
 
         String directoryBody = "<!DOCTYPE html>\n" +
                 "<html lang=\"en\">\n" +
@@ -42,6 +42,6 @@ public class DirectoryCreatorTest {
                 "</div></body>\n" +
                 "</html>";
 
-        assertEquals(directoryBody, directoryCreator.generateHTML());
+        assertEquals(directoryBody, directoryPageCreator.generateHTML());
     }
 }
